requirements.txt
boto3
Flask==2.3.2
flask-cors==4.0.0
python-dotenv==1.0.0
Pillow>=8.1.1
urllib3<1.26
gunicorn==20.1.0
redis==5.0.1
